Change things
Right. So this is a horrible commit. The result of doing an hour here
and there after work and never committing as each task got finished.

I also realized while going through this that I had been just coding
away without thinking. And that many of my pieces didn't fit well. When
going from the outside in it started to get better but that lead to
refactors and changes. And all of them are probably not done yet.

Just deciding to push _something_ now and just take the pain of bad
practices and do better going forward.

Anyone reading this because of some terrible unexplained hunk of code,
I'm sorry. I'll do better. :)