Parallelize the running commands
This turns out to be a lot easier than I thought since we're just
spinning out to the OS to handle the actual running of the command.

So to test that it's being done, just assume that the tests that run
commands continue to work. And that mocking everything out and seeing
that the concurrent.futures library is being used as intended and we
should be fine as far as parallelzing goes.

By adding some sleeps to the test commands and running in parallel I
saw it running with the expected total run time, so it's working! :D